pyproject.toml
[project]
name = "home-server-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
]
